pyproject.toml
[build-system]
requires = ["hatchling"]
build-backend = "hatchling.build"

[tool.hatch.version]
path = "azure_helper/__init__.py"

[publish.index.repos]
main = "https://upload.pypi.org/legacy/"
test = "https://test.pypi.org/legacy/"

[project]
name = "azure_mlops_helper"
authors = [{name = "Mathieu Klimczak", email = "klimczak.mathieu@pm.me"}]
readme = "README.md"
license = {file = "LICENSE.txt"}
classifiers = ["License :: OSI Approved :: MIT License"]
dynamic = ["version"]
description = "Azure helper for e2e ml pipelines"
requires-python = ">=3.8"
dependencies = [
    "azure-storage-blob>=12.13.1",
    "azureml-sdk>=1.44.0",
    "dynaconf>=3.1.9",
    "joblib>=1.1.0",
    "loguru>=0.6.0",
    "pandas>=1.4.3",
    "requests>=2.28.1",
    "scikit-learn>=1.1.1",
    "skl2onnx>=1.12"
]


[project.optional-dependencies]
dev=[
    "black>=22.6.0",
    "darglint>=1.8.1",
    "flake8>=4.0.1",
    "isort>=5.10.1",
    "mypy>=0.961",
    "pre-commit>=2.20.0",
    "radon>=5.1.0",
    "rope>=1.2.0",
    "wemake-python-styleguide>=0.16.1",
    "hatch>=1.3.1",
]

test = [
    "codecov>=2.1.12",
    "pytest>=7.1.2",
    "pytest-cov>=3.0.0",
    "pytest-xdist>=2.5.0",
    "pytest-mock>=3.8.2",
    "tox>=3.25.1",
]

doc = [
    "mike>=1.1.2",
    "mkdocs>=1.3.0",
    "mkdocs-material>=8.3.9",
    "mkdocstrings>=0.19.0",
    "mkdocstrings-python>=0.7.1",
]

[project.urls]
Home = "https://dev.azure.com/klimczakmathieu/test/_git/azure-helper-core"


[tool.pytest.ini_options]
# search for tests only in tests/ directory
testpaths = ["tests"]
# make XPASS (“unexpectedly passing”) result to fail the test suite
xfail_strict = true
addopts = "-v -n auto -l -rsxX --strict-markers --tb=short --doctest-modules --cov --cov-append --cov-report=xml:coverage.xml --cov-report=html:htmlcov --junitxml=unit-testresults.xml"

[tool.isort]
multi_line_output = 3
include_trailing_comma = true
force_grid_wrap = 0
use_parentheses = true
ensure_newline_before_comments = true
line_length = 88
